Bugfix: allow multiple objects of TermTransformer
The initialization of ttft must not be static if there is more than one object of a class derived from TermTransformer. Otherwise, its constructor argument (*this) might be freed by the time the second object is initialized, and in such case the vtable of the second object will be broken and calls to transformSubterm (virtual metod) cause segfaults. This bug occurred when the preprocessing (Kernel/Preprocessing.cpp) began to normalize the input problem twice - before and after adding theory axioms. InterpretedNormalizer creates objects of the class NLiteralTransformer (derived from TermTransformer). I spent two days hunting down this bug.