improve output format for clauses
Recently there were several additions to the saturation output,
which all can be seen as some form of statistics:
- age a, weight w, optionally number of selected literals s,
  the fact that the clause is ground;
  with syntax (a:w) / (a:w:s) / (a:w:G) / (a:w:s:G)
- optionally the fact that a clause is derived only from
  theory-axioms, with syntax T
- some information i regarding induction, with syntax I(i)

This commit puts all of these into a single statistics object
of the form {key1:value1,...,keyN:valueN}, which is placed at
the end of the line.
The new format is more uniform and therefore easier to parse,
in particular for automated tools like the visualization.
Furthermore, it can easily be extended in the future by adding
new key-value pairs.

The statistics are only printed, if proof_extra is different from
OFF. This allows us to use the same code for printing saturation
information and for printing (the units which are clauses in) the
proof.