Extend/reorganize MLMatcher
Now supports
- querying which alternatives were matched
- querying the variable bindings used in the matching
- enumerating multiple matches

Furthermore, MLMatcher is now a proper class and stores the match data
in member variables instead of global variables.

This change is mostly about reorganizing the existing code and
variables.  The matching algorithm itself has not been changed
(besides allowing to continue after the first match).
Extraction of matched alternatives and variable bindings uses the
existing data structures, and does not incur any overhead for the
existing uses of MLMatcher.