implement interactive clause selection
Adds an option manual_cs (with default off) to
perform clause selection interactively.
If manual_cs is set to on, then whenever a
clause has to be selected, the user is asked
to enter the id of a clause currently in the
set of passive clauses. The corresponding
clause then gets activated.
While this mechanism can be used on
its own, it is designed to be used in
conjunction with the Saturation Visualization.