improve exhaustiveness theory axiom
This commit changes the internally added theory axiom for exhaustiveness
of term algebras, by generating a clause instead of a formula in the
case where no boolean-valued destructors occur as part of the
term-algebra.
We would like to add the axiom as a clause instead of a
formula in all cases, but it is unclear how to do this in an optimal way
(the subtle difficulties are documented at the beginning of the method
which adds the exhaustiveness axiom).