refactor running sums for theory-clause-selection - Part 5
This commit refactors running sums for theory-clause-selection into
the Inference-class, by building on the previous commits.
As a result of this commit, it is now simple to compute the running
sums without generating the actual clause of the inference. This allows
us to use the values of the running sums in the LRS-optimizations in
BinaryResolution and Superposition.