refactor and fix computation of LRS-limits in Superposition
This commit refactors the limit-checks in Superposition to use the
new interface of the Limits-class, analogously to the previous commit
for BinaryResolution. As a side-effect, it fixes the following problem:
While the limit-checks correctly estimate the weight of the currently
built clause, they don't correctly compute the weightForClauseSelection
from the weight. While the nonGoalWeightCoefficient is factored in,
other factors are not (e.g. multiplying the weight by 2 if "inw" is on).