refactor, fix, and document usage of LRS-limits in AWPassiveClauseContainer
This commit refactors the limit-checks in AWPassiveClauseContainer to
use the new interface of the Limits-class, in a similar way to the
previous two commits. As a side-effect, it fixes the following problem:
While the limit-checks correctly estimate the weight of the currently
built clause, they don't correctly compute the weightForClauseSelection
from the weight. While the nonGoalWeightCoefficient is factored in,
other factors are not (e.g. multiplying the weight by 2 if "inw" is on).