fix Makefile
fix bug: linking of Vampire crashes for 'make vampire_rel' and
'make vampire_dbg', if called during rebasing. The problem is that the
computation of the branch-name in such a case does not work and returns
'(no' as branch-name. This commit fixes this case and makes sure to also
preserve existing behaviour for the other special case of building
Vampire while being in detached state.