remove incorrect removal of clauses from ActiveClauseContainer
This commit removes the following wrong behaviour: after concluding
using LRS that a clause in the ActiveClauseContainer (sic!) will not be
selected, we remove it from the ActiveClauseContainer.

It seems that adding this behaviour was a copy-and-paste error, which
was made a long time ago.

Note that similar reasoning makes sense for a PassiveClauseContainer,
but definitely not for an ActiveClauseContainer (since the clauses in
the ActiveClauseContainer have already been selected and don't need to
be selected to contribute further to the search space).

Note further that we still keep the behaviour to remove a clause from
the ActiveClauseContainer, if we can conclude using LRS that no child
of that clause will be selected.