MultiSplitQueue: improve interface for subclassing
This commit changes the pure virtual functions which need to be
implemented in a subclass of PredicateSplitPassiveClauseContainer.
In particular, two methods need to be implemented:
1) evaluateFeature(Clause* cl), which evaluates the feature
on an already existing clause (this method is called for LRS-computation
in Passive- and Active-clause-containers)
2) evaluateFeatureEstimate(Inference* inf), which evaluates the feature
on a clause which has not been fully constructed yet (this method is
called during LRS-optimizations in Resolution and Superposition).
This change enables more control in the case where evaluating the feature
is expensive (e.g. if we use a neural network for evaluation): in such
a case, the subclass can skip the expensive evaluation in
evaluateFeatureEstimate() by returning a (conservative) value which
ensures that the clause is classified as belonging to all queues nested
in the PredicateSplitPassiveClauseContainer.