fix -tha some option value for Nat
This commit refactors the theory axioms for natural numbers, so that
they are included not only if -tha is set to on, but also for -tha some.

It is motivated by the idea that we sometimes want to exclude expensive
integer-theory axioms (by setting -tha to some). In that case, we still
want to include all theory axioms for natural numbers, since they are
much more likely than integer axioms to contribute to a proof in the
setting of Rapid.