extend MultiSplitQueues to support Tammet-style arrangement
This commit extends the implementation of
PredicateSplitPassiveClauseContainer with a flag _layeredArrangement,
which can be configured by subclasses. If the flag is set to true, the
behaviour of the existing implementation is preserved. If the flag is
set to false, a clause is only inserted into a single queue (into the
best queue as computed by the used clause-feature), mimicking the
behaviour of layered queues as described for a special case by Tammet.

This commit furthermore adds new (experimental) boolean options -thsql,
-avsql and -slsql, to control whether a layered arrangement should be
used for each implemented instantiation of multi-split-queues.