add implementation for positive-literal multi-split-queue
This commit instantiates multi-split-queues with a new feature, which
measures the number of positive literals in the clause.

By default, the new implementation is off. It can be turned on using
using '-plsq on'. Analogously to other features
instantiating multi-split-queues, one can further tune the
implementation using the experimental options 'plsqc','plsqr',
and 'plsql'.

Motivation:
Many problems, in particular those coming from applications, contain
only a few "case distinctions", where a "case distinction" roughly
amounts to an inference, where one parent is an (input-)clause with more
than one positive literal. While this characterization is not completely
accurate, as e.g. negated disequalities are represented as positive
equalities, it still is a valuable estimate.
The idea now is to prefer the selection of clauses which have been
derived using few "case distinctions", or in other words, to prefer the
selection of clauses which have only have few positive literals.