remove Unit::getPriority() and env.clausePriority
This commit applies changes originating in a discussion
with Giles and Martin, and is motivated by getting rid
of the unmaintained clause priorities, initially hacked for LTB
several years ago.

The commit removes the mutually dependent Unit::getPriority()
and env.clausePriority and their usage in Clause::weight().
As clause priorities are removed, also the value PRIORITY for
SineSelection, which adapts priorities according to sine-levels,
is removed.

As a side-effect of this commit, we get a step closer towards
Clause::weight() returning exactly the sum of the weights of the
literals of the clause (the only remaining deviation is that size
of the split-set is still conditionally added).

As another side-effect, we remove a memory-leak:
env.clausePriority was never deleted.