extend SMTLIB parser to support "assert-theory"
This commit extends the SMTLIB parser to support a new command
"assert-theory", which is a custom extension of the SMTLIB-syntax.

It can be used as (assert-theory F), for an arbitrary formula F.

It behaves the same as (assert F), with the difference that F will be
treated as external theory axiom instead of being treated as an assumption.
This high-level information allows Vampire to treat F differently from
other axioms, e.g. to use theory-split-queues to limit the blowup caused
by F.