fix LRS-simulation for -awr 0:n and -awr n:0
The current implementation supports clause-selection using only
age or weight as criterion, by setting an awr-value n:0 resp. 0:n.

The recent implementation of layered-clause selection did not support
this corner case yet in its LRS-simulation implementation, leading to
assertion-violations in debug-mode and incorrect clause selection in
release-mode for this corner case.

This commit fixes these problems, by adapting the LRS-simulation methods.

Furthermore, it adapts ageLimited() and weightLimited() to only return
true, if the age- resp. weight-queue is actually used. As a result,
we are able to conclude more often, that we saturated the set of input
without deleting non-redundant clauses.

As a future change, it could make sense to modularize the implementation
of the age-weight-container into an age-container, a weight-container,
and a round-robin container, which has the age-container and the
weight-container as child-containers, to which it alternatingly delegates
clause-selection.
We could then use the age-container and the weight-container directly to
implement clause selection using age resp. weight as only criterion,
instead of the current implementation which in a nutshell hacks a
round-robin container to ignore one of its child-containers.
Such an implementation would have avoided the current bug, and should
be easier to maintain in general.